wtf, this stopped working for me? Why?

+++ frontend/src/i18n.js
@@ -70,7 +70,7 @@ const dateTimeFormats = {
     long: {
       year: 'numeric',
       month: 'long',
-      day: 'long',
+      day: 'numeric',
       weekday: 'short',
       hour: 'numeric',
       minute: 'numeric',
